Add file-scoped namespace settings
