--- src/services/relationship/Twitter.Clone.Relationship.APIs/README.md
@@ -1,13 +1,0 @@
-# Twitter relationship service
-
-This service's main purpose is to make the follow relationship between users, as followers and followings.
-
-## Features:
-1. Make connection between users
-2. Show the fetched data of individual users (such as followers, followings and count of them)
-3. Notify users when they get new followers
-4. Etc.
-
-
-** Development team: **
-[InfiNetCore](https://github.com/orgs/devblogs-ir/teams/infinetcore)
