+++ src/services/relationship/Twitter.Clone.Relationship/Twitter.Clone.Relationship.sln
@@ -5,6 +5,11 @@ VisualStudioVersion = 17.8.34330.188
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Twitter.Clone.Relationship.APIs", "Twitter.Clone.Relationship.APIs.csproj", "{AA848F35-5B78-4CE0-97DA-4294FE6CA59B}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{764A5ECD-9E16-4D7E-B779-CAFC15972B62}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
